compte searching funcitonality

+++ src/Components/MyToys/MyToys.jsx
@@ -3,6 +3,7 @@ import { AuthContext } from '../../Context/AuthProvider';
 import MyToysCard from '../MyToysCard';
 import Loadings from '../../Pages/Shared/Loadings.jsx'
 import useDynamicTitle from '../../Hooks/useHook';
+import Swal from "sweetalert2";
 const MyToys = () => {
   useDynamicTitle('My Toys');
 
@@ -23,19 +24,36 @@ const MyToys = () => {
 
 
       
-      
-      const handleDelete = (id) => {
 
-        fetch(`http://localhost:5000/toys/${id}`, {
-          method: 'DELETE',
-        })
-        .then(res => res.json())
-        .then(data => {
-          if(data.deletedCount > 0 ){
-            const remaining = toys.filter(toy => toy._id !== id)
-            setToys(remaining)
-          }
-        })
+
+     
+      /* ------------------------ DELETE USER DATA FUNCTION ----------------------- */
+      const handleDelete = (id) => {
+         Swal.fire({
+           title: "Are you sure?",
+           text: "You won't be able to revert this!",
+           icon: "warning",
+           showCancelButton: true,
+           confirmButtonColor: "#3085d6",
+           cancelButtonColor: "#d33",
+           confirmButtonText: "Yes, delete it!",
+         }).then((result) => {
+           if (result.isConfirmed) {
+            fetch(`http://localhost:5000/toys/${id}`, {
+              method: "DELETE",
+            })
+              .then((res) => res.json())
+              .then((data) => {
+                if (data.deletedCount > 0) {
+                  const remaining = toys.filter((toy) => toy._id !== id);
+                  setToys(remaining);
+                }
+              });
+             Swal.fire("Deleted!", "Your file has been deleted.", "success");
+           }
+         });
+
+        
 
       }
